refactor(storage-adapter): moved native php method proxies (apcu, file) to traits in the support namespace
Refs #34
